[CfM-R144] mausb: Align with recent change to HCD API
Commit b27f10764d800d6875fef8649099bacaba15b0bf introduced a
timeout argument to the .address_device template. This commit
aligns mausb with the new template.

BUG=None
TEST=emerge-$BOARD chromeos-kernel-5_15

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/7567967
Tested-by: Niko Tsirakis <ntsirakis@google.com>
Commit-Queue: Bruce Goodwin <bgoodwin@chromium.org>
Owners-Override: Niko Tsirakis <ntsirakis@google.com>
Auto-Submit: Niko Tsirakis <ntsirakis@google.com>

+++ drivers/usb/host/mausb/hpal_events.h
@@ -51,7 +51,7 @@ int mausb_modifyep0_event_from_user(struct mausb_device *dev,
 				    struct mausb_event *event);
 int mausb_setusbdevaddress_event_to_user(struct mausb_device *dev,
 					 u16 device_handle,
-					 u16 response_timeout);
+					 unsigned int response_timeout);
 int mausb_setusbdevaddress_event_from_user(struct mausb_device *dev,
 					   struct mausb_event *event);
 int mausb_updatedev_event_to_user(struct mausb_device *dev,
